Simple IP Grabber I made in PHP
LOOK AT README.md file for tutorial.

Don't be stupid and use this to threaten people i made this for
protection and this can be used for you want you can jack up and fiddle
around with the source code i am not 0.01% responsable if you perform a
stupid act.

- Peter aka dzt aka Virtual God